+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+# 1.0.0
 
-* TODO: Describe initial release.
+Initial release.
+
+Features:
+
+- Result type
+- Success / Error modeling
+- Functional helpers
+- Async helpers
+- High test coverage